Code reorganization (mostly moving stuff around and removing unused code), change to graph related compute functions
