+++ software/pc-emul/textualRepresentation.hpp
@@ -1,0 +1,8 @@
+#ifndef INCLUDED_TEXTUAL_REPRESENTATION
+#define INCLUDED_TEXTUAL_REPRESENTATION
+
+#include "versatPrivate.hpp"
+
+SizedString Repr(Versat* versat,FUInstance* inst,Arena* arena,bool printDelay = false);
+
+#endif // INCLUDED_TEXTUAL_REPRESENTATION
